change to 16bit node size header

+++ src/spiffs_circular_queue.h
@@ -42,7 +42,7 @@ uint8_t spiffs_circular_queue_init(circular_queue_t *cq);
  *  @param[out] elem_size   Pointer to a queue element size
  *	@return					1 on success and 0 on fail
  */
-uint8_t spiffs_circular_queue_front(const circular_queue_t *cq, uint8_t *elem, uint32_t *elem_size);
+uint8_t spiffs_circular_queue_front(const circular_queue_t *cq, uint8_t *elem, uint16_t *elem_size);
 
 /**
  *	Enqueues elem of elem_size size to the front of the queue.
@@ -53,7 +53,7 @@ uint8_t spiffs_circular_queue_front(const circular_queue_t *cq, uint8_t *elem, u
  *
  *	@return					1 on success and 0 on fail
  */
-uint8_t spiffs_circular_queue_enqueue(circular_queue_t *cq, const uint8_t *elem, const uint32_t elem_size);
+uint8_t spiffs_circular_queue_enqueue(circular_queue_t *cq, const uint8_t *elem, const uint16_t elem_size);
 
 /**
  *	Pops out the first element of the queue.
@@ -137,7 +137,7 @@ uint32_t spiffs_circular_queue_get_file_size(const circular_queue_t *cq);
  *	@param[in] cq 			    Pointer to the circular_queue_t struct
  *	@param[in] unmount_spiffs   Unmount SPIFFS on free flag
  *
- *	@return					1 when empty and 0 if not
+ *	@return					    1 on success and 0 on fail
  */
 uint8_t	spiffs_circular_queue_free(circular_queue_t *cq, const uint8_t unmount_spiffs = 1);
